move fake storages to store package
